+++ libs/request/src/CommonMediaResponse.ts
@@ -1,6 +1,5 @@
 import type { Response } from '@svta/cml-utils'
 import type { CommonMediaRequest } from './CommonMediaRequest.ts'
-import type { ResourceTiming } from './ResourceTiming.ts'
 
 /**
  * Common response API.
@@ -8,10 +7,4 @@ import type { ResourceTiming } from './ResourceTiming.ts'
  * @public
  */
 
-export type CommonMediaResponse<R extends CommonMediaRequest = CommonMediaRequest> =
-	Response<R> & {
-	/**
-	 * The network timing of the request/response.
-	 */
-	resourceTiming?: ResourceTiming;
-};
+export type CommonMediaResponse<R extends CommonMediaRequest = CommonMediaRequest> = Response<R>;